Create ExternalInformationSkill
