+++ samples/apps/copilot-chat-app/webapi/CopilotChat/Skills/ChatSkills/ChatSkill.cs
@@ -1,24 +1,18 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 
 using System;
-using System.Collections.Generic;
 using System.Globalization;
 using System.Linq;
 using System.Text.Json;
-using System.Text.Json.Nodes;
-using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using Microsoft.SemanticKernel;
 using Microsoft.SemanticKernel.AI.TextCompletion;
 using Microsoft.SemanticKernel.Orchestration;
-using Microsoft.SemanticKernel.Planning;
 using Microsoft.SemanticKernel.SkillDefinition;
 using SemanticKernel.Service.CopilotChat.Models;
 using SemanticKernel.Service.CopilotChat.Options;
-using SemanticKernel.Service.CopilotChat.Skills.OpenApiSkills.GitHubSkill.Model;
-using SemanticKernel.Service.CopilotChat.Skills.OpenApiSkills.JiraSkill.Model;
 using SemanticKernel.Service.CopilotChat.Storage;
 
 namespace SemanticKernel.Service.CopilotChat.Skills.ChatSkills;
@@ -29,11 +23,6 @@ namespace SemanticKernel.Service.CopilotChat.Skills.ChatSkills;
 /// </summary>
 public class ChatSkill
 {
-    /// <summary>
-    /// Logger
-    /// </summary>
-    private readonly ILogger _logger;
-
     /// <summary>
     /// A kernel instance to create a completion function since each invocation
     /// of the <see cref="ChatAsync"/> function will generate a new prompt dynamically.
@@ -56,16 +45,6 @@ public class ChatSkill
     private readonly PromptsOptions _promptOptions;
 
     /// <summary>
-    /// CopilotChat's planner to gather additional information for the chat context.
-    /// </summary>
-    private readonly CopilotChatPlanner _planner;
-
-    /// <summary>
-    /// Proposed plan to return for approval.
-    /// </summary>
-    private Plan? _proposedPlan;
-
-    /// <summary>
     /// Create a new instance of <see cref="ChatSkill"/>.
     /// </summary>
     public ChatSkill(
@@ -76,12 +55,10 @@ public class ChatSkill
         CopilotChatPlanner planner,
         ILogger logger)
     {
-        this._logger = logger;
         this._kernel = kernel;
         this._chatMessageRepository = chatMessageRepository;
         this._chatSessionRepository = chatSessionRepository;
         this._promptOptions = promptOptions.Value;
-        this._planner = planner;
     }
 
     /// <summary>
@@ -129,93 +106,6 @@ public class ChatSkill
         }
 
         return $"User intent: {result}";
-    }
-
-    /// <summary>
-    /// Extract relevant additional knowledge using a planner.
-    /// </summary>
-    [SKFunction("Acquire external information")]
-    [SKFunctionName("AcquireExternalInformation")]
-    [SKFunctionContextParameter(Name = "userIntent", Description = "The intent of the user.")]
-    [SKFunctionContextParameter(Name = "tokenLimit", Description = "Maximum number of tokens")]
-    public async Task<string> AcquireExternalInformationAsync(SKContext context)
-    {
-        FunctionsView functions = this._planner.Kernel.Skills.GetFunctionsView(true, true);
-        if (functions.NativeFunctions.IsEmpty && functions.SemanticFunctions.IsEmpty)
-        {
-            return string.Empty;
-        }
-
-        // Check if plan exists in ask's context variables.
-        // If plan was returned at this point, that means it was approved and should be run
-        var planApproved = context.Variables.Get("proposedPlan", out var planJson);
-
-        if (planApproved)
-        {
-            // Reload the plan with the planner's kernel so
-            // it has full context to be executed
-            var newPlanContext = new SKContext(
-                null,
-                this._planner.Kernel.Memory,
-                this._planner.Kernel.Skills,
-                this._planner.Kernel.Log
-            );
-            var plan = Plan.FromJson(planJson, newPlanContext);
-
-            // Invoke plan
-            newPlanContext = await plan.InvokeAsync(newPlanContext);
-            int tokenLimit = int.Parse(context["tokenLimit"], new NumberFormatInfo());
-
-            // The result of the plan may be from an OpenAPI skill. Attempt to extract JSON from the response.
-            bool extractJsonFromOpenApi = this.TryExtractJsonFromOpenApiPlanResult(newPlanContext.Result, out string planResult);
-            if (extractJsonFromOpenApi)
-            {
-                int relatedInformationTokenLimit = (int)Math.Floor(tokenLimit * this._promptOptions.RelatedInformationContextWeight);
-                planResult = this.OptimizeOpenApiSkillJson(planResult, relatedInformationTokenLimit, plan);
-            }
-            else
-            {
-                // If not, use result of the plan execution result directly.
-                planResult = newPlanContext.Variables.Input;
-            }
-
-            string informationText = $"[START RELATED INFORMATION]\n{planResult.Trim()}\n[END RELATED INFORMATION]\n";
-
-            // Adjust the token limit using the number of tokens in the information text.
-            tokenLimit -= Utilities.TokenCount(informationText);
-            context.Variables.Set("tokenLimit", tokenLimit.ToString(new NumberFormatInfo()));
-
-            return informationText;
-        }
-        else
-        {
-            // Create a plan and set it in context for approval.
-            Plan plan = await this._planner.CreatePlanAsync(context["userIntent"]);
-
-            if (plan.Steps.Count > 0)
-            {
-                // Merge any variables from the context into plan's state 
-                // as these will be used on plan execution.
-                // These context variables come from user input, so they are prioritized.
-                var variables = context.Variables;
-                foreach (var param in plan.State)
-                {
-                    if (param.Key.Equals("INPUT", StringComparison.OrdinalIgnoreCase))
-                    {
-                        continue;
-                    }
-
-                    if (variables.Get(param.Key, out var value))
-                    {
-                        plan.State.Set(param.Key, value);
-                    }
-                }
-
-                this._proposedPlan = plan;
-            }
-        }
-
-        return string.Empty;
     }
 
     /// <summary>
@@ -393,181 +283,6 @@ public class ChatSkill
     }
 
     #region Private
-
-    /// <summary>
-    /// Try to extract json from the planner response as if it were from an OpenAPI skill.
-    /// </summary>
-    private bool TryExtractJsonFromOpenApiPlanResult(string openApiSkillResponse, out string json)
-    {
-        try
-        {
-            JsonNode? jsonNode = JsonNode.Parse(openApiSkillResponse);
-            string contentType = jsonNode?["contentType"]?.ToString() ?? string.Empty;
-            if (contentType.StartsWith("application/json", StringComparison.InvariantCultureIgnoreCase))
-            {
-                var content = jsonNode?["content"]?.ToString() ?? string.Empty;
-                if (!string.IsNullOrWhiteSpace(content))
-                {
-                    json = content;
-                    return true;
-                }
-            }
-        }
-        catch (JsonException)
-        {
-            this._logger.LogDebug("Unable to extract JSON from planner response, it is likely not from an OpenAPI skill.");
-        }
-        catch (InvalidOperationException)
-        {
-            this._logger.LogDebug("Unable to extract JSON from planner response, it may already be proper JSON.");
-        }
-
-        json = string.Empty;
-        return false;
-    }
-
-    /// <summary>
-    /// Try to optimize json from the planner response
-    /// based on token limit
-    /// </summary>
-    private string OptimizeOpenApiSkillJson(string jsonContent, int tokenLimit, Plan plan)
-    {
-        int jsonTokenLimit = (int)(tokenLimit * this._promptOptions.RelatedInformationContextWeight);
-
-        // Remove all new line characters + leading and trailing white space
-        jsonContent = Regex.Replace(jsonContent.Trim(), @"[\n\r]", string.Empty);
-        var document = JsonDocument.Parse(jsonContent);
-        string lastSkillInvoked = plan.Steps[^1].SkillName;
-        string lastSkillFunctionInvoked = plan.Steps[^1].Name;
-        bool trimSkillResponse = false;
-
-        // The json will be deserialized based on the response type of the particular operation that was last invoked by the planner
-        // The response type can be a custom trimmed down json structure, which is useful in staying within the token limit
-        Type skillResponseType = this.GetOpenApiSkillResponseType(ref document, ref lastSkillInvoked, ref lastSkillFunctionInvoked, ref trimSkillResponse);
-
-        if (trimSkillResponse)
-        {
-            // Deserializing limits the json content to only the fields defined in the respective OpenApiSkill's Model classes
-            var skillResponse = JsonSerializer.Deserialize(jsonContent, skillResponseType);
-            jsonContent = skillResponse != null ? JsonSerializer.Serialize(skillResponse) : string.Empty;
-            document = JsonDocument.Parse(jsonContent);
-        }
-
-        int jsonContentTokenCount = Utilities.TokenCount(jsonContent);
-
-        // Return the JSON content if it does not exceed the token limit
-        if (jsonContentTokenCount < jsonTokenLimit)
-        {
-            return jsonContent;
-        }
-
-        List<object> itemList = new();
-
-        // Some APIs will return a JSON response with one property key representing an embedded answer.
-        // Extract this value for further processing
-        string resultsDescriptor = "";
-
-        if (document.RootElement.ValueKind == JsonValueKind.Object)
-        {
-            int propertyCount = 0;
-            foreach (JsonProperty property in document.RootElement.EnumerateObject())
-            {
-                propertyCount++;
-            }
-
-            if (propertyCount == 1)
-            {
-                // Save property name for result interpolation
-                JsonProperty firstProperty = document.RootElement.EnumerateObject().First();
-                jsonTokenLimit -= Utilities.TokenCount(firstProperty.Name);
-                resultsDescriptor = string.Format(CultureInfo.InvariantCulture, "{0}: ", firstProperty.Name);
-
-                // Extract object to be truncated
-                JsonElement value = firstProperty.Value;
-                document = JsonDocument.Parse(value.GetRawText());
-            }
-        }
-
-        // Detail Object
-        // To stay within token limits, attempt to truncate the list of properties
-        if (document.RootElement.ValueKind == JsonValueKind.Object)
-        {
-            foreach (JsonProperty property in document.RootElement.EnumerateObject())
-            {
-                int propertyTokenCount = Utilities.TokenCount(property.ToString());
-
-                if (jsonTokenLimit - propertyTokenCount > 0)
-                {
-                    itemList.Add(property);
-                    jsonTokenLimit -= propertyTokenCount;
-                }
-                else
-                {
-                    break;
-                }
-            }
-        }
-
-        // Summary (List) Object
-        // To stay within token limits, attempt to truncate the list of results
-        if (document.RootElement.ValueKind == JsonValueKind.Array)
-        {
-            foreach (JsonElement item in document.RootElement.EnumerateArray())
-            {
-                int itemTokenCount = Utilities.TokenCount(item.ToString());
-
-                if (jsonTokenLimit - itemTokenCount > 0)
-                {
-                    itemList.Add(item);
-                    jsonTokenLimit -= itemTokenCount;
-                }
-                else
-                {
-                    break;
-                }
-            }
-        }
-
-        return itemList.Count > 0
-            ? string.Format(CultureInfo.InvariantCulture, "{0}{1}", resultsDescriptor, JsonSerializer.Serialize(itemList))
-            : string.Format(CultureInfo.InvariantCulture, "JSON response for {0} is too large to be consumed at this time.", lastSkillInvoked);
-    }
-
-    private Type GetOpenApiSkillResponseType(ref JsonDocument document, ref string lastSkillInvoked, ref string lastSkillFunctionInvoked, ref bool trimSkillResponse)
-    {
-        Type skillResponseType = typeof(object); // Use a reasonable default response type
-
-        // Different operations under the skill will return responses as json structures;
-        // Prune each operation response according to the most important/contextual fields only to avoid going over the token limit
-        // Check what the last skill invoked was and deserialize the JSON content accordingly
-        if (string.Equals(lastSkillInvoked, "GitHubSkill", StringComparison.Ordinal))
-        {
-            trimSkillResponse = true;
-            skillResponseType = this.GetGithubSkillResponseType(ref document);
-        }
-        else if (string.Equals(lastSkillInvoked, "JiraSkill", StringComparison.Ordinal))
-        {
-            trimSkillResponse = true;
-            skillResponseType = this.GetJiraSkillResponseType(ref document, ref lastSkillFunctionInvoked);
-        }
-
-        return skillResponseType;
-    }
-
-    private Type GetGithubSkillResponseType(ref JsonDocument document)
-    {
-        return document.RootElement.ValueKind == JsonValueKind.Array ? typeof(PullRequest[]) : typeof(PullRequest);
-    }
-
-    private Type GetJiraSkillResponseType(ref JsonDocument document, ref string lastSkillFunctionInvoked)
-    {
-        if (lastSkillFunctionInvoked == "GetIssue")
-        {
-            return document.RootElement.ValueKind == JsonValueKind.Array ? typeof(IssueResponse[]) : typeof(IssueResponse);
-        }
-
-        return typeof(IssueResponse);
-    }
 
     /// <summary>
     /// Save a new message to the chat history.
